define class IkFastSolver<T> in header; include it in ikfastsolver.cpp
